+++ src/features/create-post/ui/crop-photo/CropPhotoStep.tsx
@@ -10,9 +10,9 @@ import {
   type CreatePostImageSize,
   MAX_CREATE_POST_ZOOM,
   MIN_CREATE_POST_ZOOM,
-} from '../model/createPostCrop'
-import type { CreatePostPhoto } from '../model/createPostFile'
-import styles from './createPost.module.css'
+} from '../../model/createPostCrop'
+import type { CreatePostPhoto } from '../../model/createPostFile'
+import styles from '../createPost.module.css'
 import { SelectedPhotosList } from './SelectedPhotosList'
 
 type Props = {