+++ src/features/create-post/ui/publication/PhotoCarouselControls.tsx
@@ -2,8 +2,8 @@ import clsx from 'clsx'
 
 import { Icon } from '@/shared/ui/icon'
 
-import type { CreatePostPhoto } from '../model/createPostFile'
-import styles from './createPost.module.css'
+import type { CreatePostPhoto } from '../../model/createPostFile'
+import styles from '../createPost.module.css'
 
 type Props = {
   photos: CreatePostPhoto[]